requirements.txt
fastapi
fastapi-cache2
uvicorn
httpx
beautifulsoup4
lxml
pydantic
